requirements.txt
streamlit 
langchain 
langchain_ollama
selenium
beautifulsoup4
lxml 
html5lib
python-dotenv
bs4
requests
langchain_google_genai
google.generativeai
transformers
aiofiles
websocket-client
gunicorn
flask
python-dotenv